fix(#7581): move Rollback into its own file and satisfy qulice
