+++ eo-parser/src/main/java/org/eolang/parser/Eo.java
@@ -279,7 +279,7 @@ final class Eo implements Iterable<Directive> {
     ) {
         if (Eo.closesTextBlock(span, globals)) {
             stack.popDeeperThan(span.indent());
-            final Rollback point = new Rollback(stack, emit);
+            final Rollback point = Rollback.of(stack, emit);
             try {
                 new LnTextBlock(span).into(stack, globals, emit);
             } catch (final ParseError err) {
@@ -327,7 +327,7 @@ final class Eo implements Iterable<Directive> {
         if (!span.blank() && span.head() != '#') {
             stack.popDeeperThan(span.indent());
         }
-        final Rollback point = new Rollback(stack, emit);
+        final Rollback point = Rollback.of(stack, emit);
         final Globals saved = globals.savepoint();
         boolean failed = false;
         try {
@@ -677,55 +677,5 @@ final class Eo implements Iterable<Directive> {
             parent.openTuple();
         }
         parent.child();
-    }
-
-    /**
-     * A joint savepoint over the {@link Emit} sink and the {@link Stack}
-     * levels, taken before a recoverable parse and rolled back together
-     * on a {@link ParseError} (§7).
-     *
-     * @since 0.1
-     */
-    private static final class Rollback {
-
-        /**
-         * The stack to restore on {@link #apply()}.
-         */
-        private final Stack stack;
-
-        /**
-         * The sink to roll back on {@link #apply()}.
-         */
-        private final Emit emit;
-
-        /**
-         * The sink savepoint taken at construction.
-         */
-        private final int token;
-
-        /**
-         * The stack levels snapshot taken at construction.
-         */
-        private final java.util.List<Level> frame;
-
-        /**
-         * Ctor.
-         * @param stack The stack to snapshot and later restore
-         * @param emit The sink to mark and later roll back
-         */
-        Rollback(final Stack stack, final Emit emit) {
-            this.stack = stack;
-            this.emit = emit;
-            this.token = emit.savepoint();
-            this.frame = stack.snapshot();
-        }
-
-        /**
-         * Roll the sink and the stack back to this savepoint.
-         */
-        void apply() {
-            this.emit.rollback(this.token);
-            this.stack.restore(this.frame);
-        }
     }
 }
